test: move the are-you-sure e2e test into issue-comment.test.ts
